requirements.txt
schedule
selenium
requests
python-telegram-bot